[#21] Refactor PreferenceManager
